+++ Demo/Scripts/LanguageSwitcher.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+using UnityLiteL10n;
+
+public class LanguageSwitcher : MonoBehaviour
+{
+    public void OnChangeToEN() => LocalizationManager.Instance.SetLanguage("EN");
+    public void OnChangeToFR() => LocalizationManager.Instance.SetLanguage("FR");
+    public void OnChangeToES() => LocalizationManager.Instance.SetLanguage("ES");
+}